requirements.txt
opencv-python 
matplotlib 
numpy 
Pillow 
tqdm